[Android] Remap the URL before loading in WebView
This matches the implementation on iOS and means that an incoming URL
like https://app.scope/#/foo will always cause a file:/// URL to be
loaded in the WebView.

This is useful because then we are able to rely on things like the
existence of a file:/// URL to indicate that we need to use hash-based
routing instead of pushState fragments.